+++ packages/core/src/component/component-name.ts
@@ -1,4 +1,5 @@
 import { getFunctionIdentifier, type TSESTreeFunction } from "@eslint-react/ast";
+import { F, O } from "@eslint-react/tools";
 import type { TSESTree } from "@typescript-eslint/types";
 
 export const RE_COMPONENT_NAME = /^[A-Z]/u;
@@ -14,7 +15,11 @@ export function isValidComponentName(name: null | string | undefined): name is s
 }
 
 export function hasNoneOrValidComponentName(node: TSESTreeFunction) {
-  const id = getFunctionIdentifier(node);
-
-  return !id || isValidComponentName(id.name);
+  return O.match(
+    getFunctionIdentifier(node),
+    {
+      onNone: F.constTrue,
+      onSome: id => isValidComponentName(id.name),
+    },
+  );
 }